feat: refactors attention mechanism
