+++ lzero/model/unizero_world_models/modeling/__init__.py
@@ -1,0 +1,3 @@
+"""
+Attention module for the Transformer backbone of UniZero.
+"""
