+++ lzero/model/unizero_world_models/modeling/global_attention.py
@@ -1,0 +1,167 @@
+"""
+Standard Self-Attention Module for Transformers (Global Attention)
+"""
+import math
+from typing import Optional
+
+import torch
+import torch.nn as nn
+from torch.nn import functional as F
+from einops import rearrange
+from .kv_caching import KeysValues
+from .transformer import apply_rotary_emb, TransformerConfig
+from .attention import Attention
+
+
+class GlobalAttention(Attention):
+    """
+    Implements self-attention mechanism for transformers.
+
+    Arguments:
+        config (:obj:`TransformerConfig`): Configuration object containing hyperparameters.
+
+    Attributes:
+        - config (:obj:`TransformerConfig`): Stores the configuration for the self-attention module.
+        - num_heads (:obj:`int`): Number of attention heads.
+        - key (:obj:`nn.Linear`): Linear layer to project input to key vectors.
+        - query (:obj:`nn.Linear`): Linear layer to project input to query vectors.
+        - value (:obj:`nn.Linear`): Linear layer to project input to value vectors.
+        - attn_drop (:obj:`nn.Dropout`): Dropout layer for attention weights.
+        - resid_drop (:obj:`nn.Dropout`): Dropout layer for residual connection.
+        - proj (:obj:`nn.Linear`): Final linear layer for projection.
+        - mask (:obj:`torch.Tensor`): Mask tensor for causal or block-causal attention.
+    """
+
+    def __init__(self, config: TransformerConfig) -> None:
+        super().__init__()
+        assert config.embed_dim % config.num_heads == 0, "Embedding dimension must be divisible by number of heads."
+
+        self.config = config
+        self.num_heads = config.num_heads
+
+        self.key = nn.Linear(config.embed_dim, config.embed_dim)
+        self.query = nn.Linear(config.embed_dim, config.embed_dim)
+        self.value = nn.Linear(config.embed_dim, config.embed_dim)
+
+        self.attn_drop = nn.Dropout(config.attn_pdrop)
+        self.resid_drop = nn.Dropout(config.resid_pdrop)
+        self.proj = nn.Linear(config.embed_dim, config.embed_dim)
+
+        causal_mask = torch.tril(torch.ones(config.max_tokens, config.max_tokens))
+        self.register_buffer('mask', causal_mask)
+
+    def forward(self, x: torch.Tensor, kv_cache: Optional[KeysValues] = None,
+                valid_context_lengths: Optional[torch.Tensor] = None, freqs_cis: torch.Tensor = None) -> torch.Tensor:
+        """
+        Forward pass for the self-attention mechanism.
+
+        Arguments:
+            - x (:obj:`torch.Tensor`): Input tensor of shape (B, T, C) where B is batch size,
+                                        T is sequence length, and C is embedding dimension.
+            - kv_cache (:obj:`Optional[KeysValues]`): Optional key-value cache for faster inference.
+            - valid_context_lengths (:obj:`Optional[torch.Tensor]`): Optional tensor containing valid context lengths.
+            - freqs_cis (:obj:`torch.Tensor`): Frequency components for rotary position embeddings, used to modulate the attention mechanism (default: None).
+
+        Returns:
+            - torch.Tensor: Output tensor of shape (B, T, C).
+        """
+        B, T, C = x.size()
+        if kv_cache is not None:
+            b, nh, L, c = kv_cache.shape
+            assert nh == self.num_heads and b == B and c * nh == C, "Cache dimensions do not match input dimensions."
+        else:
+            L = 0
+
+        q = self.query(x).view(B, T, self.num_heads, C // self.num_heads).transpose(1,
+                                                                                    2)  # (B, num_heads, T, head_size)
+        k = self.key(x).view(B, T, self.num_heads, C // self.num_heads).transpose(1, 2)  # (B, num_heads, T, head_size)
+        v = self.value(x).view(B, T, self.num_heads, C // self.num_heads).transpose(1,
+                                                                                    2)  # (B, num_heads, T, head_size)
+
+        if self.config.rotary_emb:
+            q, k = apply_rotary_emb(q, k, freqs_cis=freqs_cis)
+
+        if kv_cache is not None:
+            kv_cache.update(k, v)  # time occupancy 21%
+            k, v = kv_cache.get()  # time occupancy 5%
+
+        att = (q @ k.transpose(-2, -1)) * (1.0 / math.sqrt(k.size(-1)))
+
+        if valid_context_lengths is not None:
+            # Final mask.shape: (B, T, L + T)
+            # L is the context length, T is the current input length,
+            # valid_context_lengths is the valid length at the end of the context.
+            mask = torch.zeros(B, T, L + T, device=att.device)
+            # For each sample, set the invalid parts to 0 based on its valid length.
+            for i in range(B):
+                mask[i] = self.mask[L:L + T, :L + T].clone()
+                mask[i, :, :(L - valid_context_lengths[i])] = 0  # Set invalid parts to 0.
+                # Adjust mask dimensions to match the last two dimensions of att.
+                # (B, T, L + T) -> (B, 1, T, L + T) -> (B, num_heads, T, L + T)
+                mask = mask.unsqueeze(1).expand(-1, att.size(1), -1, -1)
+        else:
+            # mask.shape: (T, L + T)
+            mask = self.mask[L:L + T, :L + T]  # Causal Mask being Applied
+
+        # att.shape: (B, num_heads, T, L + T)
+        att = att.masked_fill(mask == 0, float('-inf'))
+
+        att = F.softmax(att, dim=-1)
+        att = self.attn_drop(att)
+        y = att @ v  # (B, num_heads, T, L + T) x (B, num_heads, L + T, head_size) -> (B, num_heads, T, head_size)
+
+        y = rearrange(y, 'b h t e -> b t (h e)')  # Combine the heads back together (B, T, embed_dim)
+        y = self.resid_drop(self.proj(y))
+
+        return y
+
+    @torch.no_grad()
+    def get_attention_map(self, x: torch.Tensor, kv_cache: Optional[KeysValues] = None,
+                          valid_context_lengths: Optional[torch.Tensor] = None) -> torch.Tensor:
+        """
+        Compute the attention map for the input sequence. This is useful for visualization purposes.
+        More details can be found in visualizing_utils.py.
+
+        Arguments:
+            - x (:obj:`torch.Tensor`): Input sequence with shape (B, T, C).
+            - kv_cache (:obj:`Optional[KeysValues]`): Cached keys and values for supporting long sequence inference.
+            - valid_context_lengths (:obj:`Optional[torch.Tensor]`): Valid context lengths for handling variable-length contexts.
+
+        Returns:
+            - torch.Tensor: Attention map with shape (B, nh, T, L + T), representing the distribution of attention.
+        """
+        B, T, C = x.size()
+        if kv_cache is not None:
+            b, nh, L, c = kv_cache.shape
+            assert nh == self.num_heads and b == B and c * nh == C, "Cache dimensions are inconsistent with input dimensions."
+        else:
+            L = 0
+
+        # Compute query, key, and value projections
+        q = self.query(x).view(B, T, self.num_heads, C // self.num_heads).transpose(1, 2)  # (B, nh, T, hs)
+        k = self.key(x).view(B, T, self.num_heads, C // self.num_heads).transpose(1, 2)  # (B, nh, T, hs)
+        v = self.value(x).view(B, T, self.num_heads, C // self.num_heads).transpose(1, 2)  # (B, nh, T, hs)
+
+        if kv_cache is not None:
+            # Update the kv_cache with the new keys and values
+            kv_cache.update(k, v)
+            k, v = kv_cache.get()
+
+        # Compute the attention scores
+        att = (q @ k.transpose(-2, -1)) * (1.0 / math.sqrt(k.size(-1)))
+
+        if valid_context_lengths is not None:
+            mask = torch.zeros(B, T, L + T, device=att.device)
+            for i in range(B):
+                # Create attention mask for each batch
+                mask[i] = self.mask[L:L + T, :L + T].clone()
+                mask[i, :, :(L - valid_context_lengths[i])] = 0
+            mask = mask.unsqueeze(1).expand(-1, att.size(1), -1, -1)
+        else:
+            mask = self.mask[L:L + T, :L + T]
+
+        # Apply the attention mask
+        att = att.masked_fill(mask == 0, float('-inf'))
+        att = F.softmax(att, dim=-1)
+
+        return att
